delete VM + VM Tests
